+++ query/transaction.go
@@ -3,7 +3,6 @@ package query
 import (
 	"context"
 
-	internal "github.com/ydb-platform/ydb-go-sdk/v3/internal/query/tx"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
 )
 
@@ -19,32 +18,32 @@ type (
 		CommitTx(ctx context.Context) (err error)
 		Rollback(ctx context.Context) (err error)
 	}
-	TransactionControl  = internal.Control
-	TransactionSettings = internal.Settings
-	TransactionOption   = internal.Option
+	TransactionControl  = tx.Control
+	TransactionSettings = tx.Settings
+	TransactionOption   = tx.Option
 )
 
 // BeginTx returns selector transaction control option
-func BeginTx(opts ...TransactionOption) internal.ControlOption {
-	return internal.BeginTx(opts...)
+func BeginTx(opts ...TransactionOption) tx.ControlOption {
+	return tx.BeginTx(opts...)
 }
 
-func WithTx(t tx.Identifier) internal.ControlOption {
-	return internal.WithTx(t)
+func WithTx(t tx.Identifier) tx.ControlOption {
+	return tx.WithTx(t)
 }
 
-func WithTxID(txID string) internal.ControlOption {
-	return internal.WithTxID(txID)
+func WithTxID(txID string) tx.ControlOption {
+	return tx.WithTxID(txID)
 }
 
 // CommitTx returns commit transaction control option
-func CommitTx() internal.ControlOption {
-	return internal.CommitTx()
+func CommitTx() tx.ControlOption {
+	return tx.CommitTx()
 }
 
 // TxControl makes transaction control from given options
-func TxControl(opts ...internal.ControlOption) *TransactionControl {
-	return internal.NewControl(opts...)
+func TxControl(opts ...tx.ControlOption) *TransactionControl {
+	return tx.NewControl(opts...)
 }
 
 func NoTx() *TransactionControl {
@@ -57,12 +56,12 @@ func DefaultTxControl() *TransactionControl {
 }
 
 // SerializableReadWriteTxControl returns transaction control with serializable read-write isolation mode
-func SerializableReadWriteTxControl(opts ...internal.ControlOption) *TransactionControl {
-	return internal.SerializableReadWriteTxControl(opts...)
+func SerializableReadWriteTxControl(opts ...tx.ControlOption) *TransactionControl {
+	return tx.SerializableReadWriteTxControl(opts...)
 }
 
 // OnlineReadOnlyTxControl returns online read-only transaction control
-func OnlineReadOnlyTxControl(opts ...internal.OnlineReadOnlyOption) *TransactionControl {
+func OnlineReadOnlyTxControl(opts ...tx.OnlineReadOnlyOption) *TransactionControl {
 	return TxControl(
 		BeginTx(WithOnlineReadOnly(opts...)),
 		CommitTx(), // open transactions not supported for OnlineReadOnly
@@ -86,30 +85,30 @@ func SnapshotReadOnlyTxControl() *TransactionControl {
 }
 
 // TxSettings returns transaction settings
-func TxSettings(opts ...internal.Option) TransactionSettings {
+func TxSettings(opts ...tx.Option) TransactionSettings {
 	return opts
 }
 
 func WithDefaultTxMode() TransactionOption {
-	return internal.WithDefaultTxMode()
+	return tx.WithDefaultTxMode()
 }
 
 func WithSerializableReadWrite() TransactionOption {
-	return internal.WithSerializableReadWrite()
+	return tx.WithSerializableReadWrite()
 }
 
 func WithSnapshotReadOnly() TransactionOption {
-	return internal.WithSnapshotReadOnly()
+	return tx.WithSnapshotReadOnly()
 }
 
 func WithStaleReadOnly() TransactionOption {
-	return internal.WithStaleReadOnly()
+	return tx.WithStaleReadOnly()
 }
 
-func WithInconsistentReads() internal.OnlineReadOnlyOption {
-	return internal.WithInconsistentReads()
+func WithInconsistentReads() tx.OnlineReadOnlyOption {
+	return tx.WithInconsistentReads()
 }
 
-func WithOnlineReadOnly(opts ...internal.OnlineReadOnlyOption) TransactionOption {
-	return internal.WithOnlineReadOnly(opts...)
+func WithOnlineReadOnly(opts ...tx.OnlineReadOnlyOption) TransactionOption {
+	return tx.WithOnlineReadOnly(opts...)
 }
